refactor: MainScreen에서 사용하는 데이터 객체를 core 패키지로 옮김
